Update README.md
# Snake-Water-Game
It is a game built with the help of fundamentals of Python Progamming language. With the help of random module, 
the computer is giving random values that is making it more fun to play with it. 
Also, concept of if, elif-else has been used to consider all the possibilities to occur.
Thus, the user-computer interaction is friendly and interactive enough.